reorganize and clean-up
