+++ src/main/resources/templates/minesweeper.html
@@ -14,7 +14,7 @@
     </form>
     <div class="info">
         <div th:text="'Status: ' + ${board.status.display}">Status: ?</div>
-        <div th:text="'Marked: ' + ${board.countMarks} + '/' + ${board.bombCount}">Marked: ?</div>
+        <div th:text="'Marked: ' + ${board.marksMade} + '/' + ${board.bombCount}">Marked: ?</div>
     </div>
 </div>
